Break out classes and add types
