+++ src/gtfs_realtime/gtfs_types.py
@@ -1,0 +1,17 @@
+from enum import Enum
+
+
+class OccupancyStatus(Enum):
+    EMPTY = 0
+    MANY_SEATS_AVAILABLE = 1
+    FEW_SEATS_AVAILABLE = 2
+    STANDING_ROOM_ONLY = 3
+    CRUSHED_STANDING_ROOM_ONLY = 4
+    FULL = 5
+    NOT_ACCEPTING_PASSENGERS = 6
+    NO_DATA_AVAILABLE = 7
+    NOT_BOARDABLE = 8
+
+
+class ScheduleRelationship(Enum):
+    pass
